+++ REcomp/common/prime_fasta_processing.py
@@ -1,8 +1,3 @@
-import os
-import re
-import shutil
-from pathlib import Path
-
 import pandas as pd
 from Bio import SeqIO
 from Bio.Blast.Applications import NcbiblastnCommandline
@@ -10,11 +5,13 @@ from Bio.Blast.Applications import NcbiblastnCommandline
 
 class FastaFinalizer:
 
-    def __init__(self, path_to_prime, path_to_final, prefix_list):
+    def __init__(self,
+                 path_to_prime,
+                 path_to_final,
+                 prefix_list):
         self.path_to_prime = path_to_prime
         self.path_to_final = path_to_final
         self.prefix_list = prefix_list
-
 
     def final_fasta(self, path_to_fasta):
         records_id = [record.id for record in SeqIO.parse(path_to_fasta,
@@ -24,7 +21,12 @@ class FastaFinalizer:
         other_id = [other for other in records_id if "Contig" in other]
         other_prefixes = [oth_id.split("_")[0] for oth_id in other_id]
         require_pref = set(other_prefixes).difference(set(rank_prefixes))
-        with open(self.path_to_final.joinpath(path_to_fasta.name), "w") as ffasta:
+        ffasta_path = self.path_to_final.joinpath(path_to_fasta.name)
+        prime_tmp_fasta_path = self.path_to_prime.joinpath(
+            f"{path_to_fasta.stem}_tmp.fasta"
+        )
+        prime_fasta_path = self.path_to_prime.joinpath(path_to_fasta.name)
+        with open(ffasta_path, "w") as ffasta:
             for record in SeqIO.parse(path_to_fasta, "fasta"):
                 if record.id not in other_id:
                     SeqIO.write(record, ffasta, "fasta")
@@ -33,17 +35,16 @@ class FastaFinalizer:
         if len(require_pref) == 0:
             return
         tmp_id = self.__tmp_other_id(require_pref, other_id)
-        with open(self.path_to_prime.joinpath(f"{path_to_fasta.stem}_tmp.fasta"),
-                  "w") as tmp, open(self.path_to_prime.joinpath(path_to_fasta.name)) as src:
+        with open(prime_tmp_fasta_path, "w") as tmp, \
+                open(prime_fasta_path) as src:
             for record in SeqIO.parse(src, "fasta"):
                 if record.id in tmp_id:
                     SeqIO.write(record, tmp, "fasta")
-        cline = NcbiblastnCommandline(query=self.path_to_final.joinpath(path_to_fasta.name),
-                                      subject=self.path_to_prime.joinpath(f"{path_to_fasta.stem}_tmp.fasta"),
+        cline = NcbiblastnCommandline(query=ffasta_path,
+                                      subject=prime_tmp_fasta_path,
                                       out="-",
                                       outfmt="6 qseqid sseqid evalue bitscore",
-                                      max_hsps=1)
-        print(cline)
+                                      task="blastn")
         output = cline()[0].strip()
         rows = [line.split() for line in output.splitlines()]
         cols = ["qseqid", "sseqid", "evalue", "bitscore"]
@@ -52,22 +53,27 @@ class FastaFinalizer:
                       "evalue": float,
                       "bitscore": float}
         b_tab = pd.DataFrame(rows, columns=cols).astype(data_types)
-        b_tab_bit = b_tab.sort_values("bitscore").iloc[[0]]["sseqid"].tolist()[0]
-        b_tab_evalue = b_tab.sort_values("evalue").iloc[[0]]["sseqid"].tolist()[0]
-        with open(self.path_to_final.joinpath(path_to_fasta.name), "a") as ffasta, \
-            open(self.path_to_prime.joinpath(f"{path_to_fasta.stem}_tmp.fasta")) as tmp:
-                for record in SeqIO.parse(tmp, "fasta"):
-                    if record.id == b_tab_bit:
-                        record.id = f"{record.id}_maxBitscore"
-                        record.description = ""
-                        SeqIO.write(record, ffasta, "fasta")
-                    elif record.id == b_tab_evalue:
-                        record.id = f"{record.id}_minEvalue"
-                        record.description = ""
-                        SeqIO.write(record, ffasta, "fasta")
-                    else:
-                        continue
-
+        if b_tab.empty:
+            return
+        b_tab_bit = (
+            b_tab.sort_values("bitscore").iloc[[0]]["sseqid"].tolist()[0]
+        )
+        b_tab_evalue = (
+            b_tab.sort_values("evalue").iloc[[0]]["sseqid"].tolist()[0]
+        )
+        with open(ffasta_path, "a") as ffasta, \
+                open(prime_tmp_fasta_path) as tmp:
+            for record in SeqIO.parse(tmp, "fasta"):
+                if record.id == b_tab_bit:
+                    record.id = f"{record.id}_maxBitscore"
+                    record.description = ""
+                    SeqIO.write(record, ffasta, "fasta")
+                elif record.id == b_tab_evalue:
+                    record.id = f"{record.id}_minEvalue"
+                    record.description = ""
+                    SeqIO.write(record, ffasta, "fasta")
+                else:
+                    continue
 
     def __tmp_other_id(self, pref_list, other_id):
         other_id_tmp = []